Add unit test that shows a potential backlog leak
The new test backlog_leak_test demonstrates the backlog counter leak which
can happen if the client process exits right after the shackle_backlog:check/3
returns true but before sending the "cast" message to the server process.

It is difficult to kill the client process at an exact time hitting such a short
time interval, so for the sake of the demo artificial delay active only in this
test was added to the shackle_pool:server/3 code (maybe_nap/1 call).